refactor(arch): move ArchSpec + ArchSpecGeometry to bloqade.lanes.arch (stage 5 of #569)
Fifth stage of the multi-stage package merge tracked in #569. Moves
the canonical home of ArchSpec and ArchSpecGeometry / BusDescriptor
from bloqade.lanes.layout.{arch,arch_geometry} to
bloqade.lanes.arch.{spec,geometry} — runtime arch surface lives in
its dedicated package alongside the builders.

Old paths bloqade.lanes.layout.arch and bloqade.lanes.layout.arch_geometry
remain as thin re-export shims, so every existing caller across
bloqade.lanes, bloqade.gemini, demos, and tests compiles unchanged.
The shims are removed in the final cleanup stage (Stage 9 of the
plan) once all in-flight branches have rebased onto the canonical
paths.

Side fix: spec.py's relative `from .word import Word` would resolve
to bloqade.lanes.arch.word (which does not exist) post-move. Switched
to an absolute import `from bloqade.lanes.layout.word import Word`,
which works on main today and continues to work via Stage 2's shim
once that ships.

Refs #569.
