+++ python/bloqade/lanes/arch/__init__.py
@@ -1,5 +1,7 @@
 from bloqade.lanes.arch.arch_builder import ArchBuilder, ZoneBuilder
 from bloqade.lanes.arch.builder import ArchResult, build_arch
+from bloqade.lanes.arch.geometry import ArchSpecGeometry, BusDescriptor
+from bloqade.lanes.arch.spec import ArchSpec
 from bloqade.lanes.arch.topology import (
     AllToAllSiteTopology,
     DiagonalWordTopology,
@@ -19,6 +21,9 @@ __all__ = [
     "ArchBlueprint",
     "ArchBuilder",
     "ArchResult",
+    "ArchSpec",
+    "ArchSpecGeometry",
+    "BusDescriptor",
     "DeviceLayout",
     "DiagonalWordTopology",
     "HypercubeSiteTopology",
